Main svc

+++ statistic/statistic-dto/src/main/java/ru/practicum/ewm/dtos/StatisticWithHitsDto.java
@@ -12,5 +12,5 @@ import lombok.Setter;
 public class StatisticWithHitsDto {
     private String app;
     private String uri;
-    private Long hits;
+    private int hits;
 }
