+++ statistic/statistic-dto/src/main/java/ru/practicum/ewm/dtos/StatisticWithHitsProjection.java
@@ -5,5 +5,5 @@ public interface StatisticWithHitsProjection {
 
     String getUri();
 
-    Long getHits();
+    int getHits();
 }
